Fixed some compiler problems
git-svn-id: https://svn.code.sf.net/p/agistudio/code/trunk/agistudio@29 97754c87-e72b-0410-a0a9-954b2e676f6f

+++ src/logcompile.cpp
@@ -755,9 +755,9 @@ void Logic::ReadArgs(bool CommandIsIf, byte CmdNum)
         }
       }// argument is inventory object and given as string
       else{ //normal argument
-        ThisArgTypePrefix = (char *)ArgTypePrefix[ThisCommand.argTypes[CurArg]];
+        ThisArgTypePrefix = (char *)ArgTypePrefix[(int)ThisCommand.argTypes[CurArg]];
         if(UseTypeChecking && (strcmp(LowerCaseArgText.substr(0,strlen(ThisArgTypePrefix)).c_str(),ThisArgTypePrefix))){
-          ShowError(CurLine,"Invalid or unknown argument type for argument "+IntToStr(CurArg)+" (should be a "+ArgTypeName[ThisCommand.argTypes[CurArg]]+").");
+          ShowError(CurLine,"Invalid or unknown argument type for argument "+IntToStr(CurArg)+" (should be a "+ArgTypeName[(int)ThisCommand.argTypes[CurArg]]+").");
         }
         else{
           if (UseTypeChecking)ArgTextPos+=strlen(ThisArgTypePrefix);
